Working on floaters. Upgraded to BEPUphysics trunk.

+++ BEPUphysics/Paths/CardinalSpline3D.cs
@@ -50,7 +50,7 @@ namespace BEPUphysics.Paths
                 Vector3 previous = ControlPoints[i - 1].Value;
                 Vector3 next = ControlPoints[i + 1].Value;
                 Vector3.Subtract(ref next, ref previous, out tangent);
-                Vector3.Multiply(ref tangent, (1 - tension) / (ControlPoints[i + 1].Time - ControlPoints[i - 1].Time), out tangent);
+                Vector3.Multiply(ref tangent, (float)((1 - tension) / (ControlPoints[i + 1].Time - ControlPoints[i - 1].Time)), out tangent);
                 tangents.Add(tangent);
             }
             tangents.Add(Vector3.Zero);
